refactor(craft): extract _ServeMixin and tighten SandboxManager after staff review
Pull the opencode-serve transport plumbing (event-bus cache, prompt slot,
readiness probe, send-message loop, subscribe stream) out of base.py into a
new sandbox/serve_transport.py module. SandboxManager composes _ServeMixin
via MRO; the abstract base is now contract-only.

Correctness fixes
- Lock-protect _init_serve_state so racing first-callers can't end up with
  different _event_buses_lock objects.
- Release per-session buses in cleanup_session_workspace via
  _close_session_buses (Docker + K8s) — previously buses leaked one reader
  thread + httpx connection per deleted session until api_server restart.
- Hoist the httpx.Client out of _wait_for_opencode_serve_ready's poll loop
  and use OpencodeServeClient.health_check() so the readiness contract has
  one implementation, not two.
- Cache per-sandbox ServeConnectionInfo (base_url + password) at first use;
  hot paths no longer re-hit docker inspect / k8s Secret reads on every
  prompt. Invalidated on terminate and re-provision.

Architecture
- send_message is concrete in base; backends implement only
  _send_message_via_acp. Identical dispatch block deleted from both
  Docker and K8s.
- Single _load_serve_connection_info abstract hook replaces the
  per-call _serve_base_url + _read_opencode_password pair (K8s keeps
  _read_opencode_password because the provision/secret-merge flow uses
  it directly).
- _close_all_sandbox_buses extracted; both terminate() impls call it
  instead of duplicating the lock-pop-close dance.
- Collapsed the duplicated GeneratorExit/Exception abort branches in
  _send_message_via_serve into _abort_and_log_turn_failure.

Cleanup
- _RenderedSessionFiles is now a NamedTuple (was a frozen dataclass with
  two fields).
- _opencode_json_write_snippet helper dedupes the two render blocks in
  Docker's setup/regen paths.
- Document the Docker plaintext OPENCODE_CONFIG_CONTENT regression vs the
  K8s Secret-backed path.
- Warn at Docker provision when all_llm_configs has >1 entry — single-
  provider Docker silently dropped overrides otherwise.
- install.sh: extract is_using_sudo() helper; the three open-coded
  ${DOCKER_SUDO[0]} == "sudo" checks now call it. Verified end-to-end
  via --no-prompt install + --shutdown (containers up & gracefully
  stopped; no sudo prefix in printed compose commands).
- SSEKeepalive re-export shim removed from base.py; three callers
  (session/manager.py + two tests) now import from sandbox.sse.

Tests
- 28 affected unit tests updated and passing (test_docker_serve_plumbing,
  test_event_bus_per_directory, test_prompt_slot now exercise the
  _ServeMixin/_load_serve_connection_info surface).
- 293 tests pass under tests/unit/onyx/server/features/build/ (the 6
  pre-existing test_send_message_with_bus.py failures from main remain).
- ruff check / ruff format --check / ty check clean across the tree.
